requirements.txt
python-telegram-bot==20.8
ccxt
ta
pandas
apscheduler
sqlalchemy
